+++ test/test-helper.js
@@ -11,6 +11,8 @@ var reactTestUtils = React.addons.TestUtils;
 var TestLocation = require('react-router/modules/locations/TestLocation');
 var Fiber = require('fibers');
 var sinon = require('sinon');
+var mockedData = require('../web/mocked-api/data/index');
+var mockedRequests = require('../web/mocked-api/requests/index');
 
 //store the original state of all the stores
 _.forEach(storeLocations, function(path, storeName) {
@@ -121,5 +123,58 @@ module.exports = {
   restoreEventHandler: function(component, eventHandlerName) {
     //using weird syntax here to prevent issue with ReactJS auto binding of events
     component.type.prototype.__reactAutoBindMap[eventHandlerName].restore();
+  },
+
+  mockedData: mockedData,
+
+  mockNockRequest: function(scope, resource, verb, key, options) {
+    var mockedRequestMetaData = mockedRequests[resource][verb][key];
+    var body = mockedRequestMetaData.requestPayload ? mockedRequestMetaData.requestPayload : undefined;
+    var mock = scope[verb](mockedRequestMetaData.url, body);
+    var replyHeaders;
+
+    if (mockedRequestMetaData.requestHeaders) {
+      _.forEach(mockedRequestMetaData.requestHeaders, function(value, header) {
+        mock.matchHeader(header, value);
+      });
+    }
+
+    if (options) {
+      if (options.responseHeaders) {
+        replyHeaders = options.responseHeaders;
+      }
+
+      if (options.times) {
+        mock.times(options.times);
+      }
+
+      if (options.delay) {
+        mock.delay = options.delay;
+      }
+    }
+
+    mock.reply(mockedRequestMetaData.httpCode, mockedRequestMetaData.response, replyHeaders);
+  },
+
+  scryRenderedDOMComponentsWithProp: function scryRenderedDOMComponentsWithProp(root, propName, propValue) {
+    return reactTestUtils.findAllInRenderedTree(root, function(inst) {
+      var instancePropValue = inst.props[propName];
+
+      return (
+        reactTestUtils.isDOMComponent(inst)
+        && instancePropValue
+        && (' ' + instancePropValue + ' ').indexOf(' ' + propValue + ' ') !== -1
+      );
+    });
+  },
+
+  findRenderedDOMComponentWithProp: function findRenderedDOMComponentWithProp(root, propName, propValue) {
+    var all = this.scryRenderedDOMComponentsWithProp(root, propName, propValue);
+
+    if (all.length !== 1) {
+      throw new Error('Did not find exactly one match (found: ' + all.length + ') for prop  ' + propName + ' : ' + propValue);
+    }
+
+    return all[0];
   }
 };